pyproject.toml
[project]
name = "lance"
version = "0.1.0"
description = "Multi-purpose Discord bot for the First Wave Survivors community"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.5",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
    "aiosqlite>=0.20",
    "dateparser>=1.2",
    "tzdata>=2024.1; sys_platform == 'win32'",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
